Another patch by  Casey Link. Better handling of mp3tunes album that has no cover art, and generally better handling of invalid cover at in services. Also, this patch was itself clean, but it triggered a pretty serious flaws in my own code ( Some confusion about wether to use names or ids as keys in collection maps ) that this commit also fixes.
svn path=/trunk/extragear/multimedia/amarok/; revision=738726

+++ src/servicebrowser/ServiceAlbumCoverDownloader.cpp
@@ -1,5 +1,6 @@
 /***************************************************************************
  *   Copyright (c) 2007  Nikolaj Hald Nielsen <nhnFreespirit@gmail.com>    *
+ *                       Casey Link <unnamedrambler@gmail.com>             *
  *                                                                         *
  *   This program is free software; you can redistribute it and/or modify  *
  *   it under the terms of the GNU General Public License as published by  *
@@ -152,8 +153,17 @@ void ServiceAlbumCoverDownloader::coverDownloadComplete( KJob * downloadJob )
     if ( downloadJob != m_albumDownloadJob )
         return ; //not the right job, so let's ignore it
 
-    m_album->setImage( QImage( m_coverDownloadPath ) );
+    QImage cover = QImage( m_coverDownloadPath );
+    if ( cover.isNull() ) {
+        debug() << "file not a valid image";
+        //the file wasn't an image, so inform album
+        m_album->imageDownloadCanceled();
 
+        return ;
+    }
+    
+    m_album->setImage( cover );
+    
     downloadJob->deleteLater();
 
     m_tempDir->unlink();
